Debug adapter sends start and stop message to webview
