+++ src/service/messagingService.ts
@@ -1,16 +1,26 @@
 import { Webview } from "vscode";
+import { VSCODE_MESSAGES_TO_WEBVIEW } from "../view/constants";
 export class MessagingService {
     private currentWebviewTarget: Webview | undefined;
 
     public setWebview(webview: Webview) {
         this.currentWebviewTarget = webview;
     }
+
+    // Send a message to webview if it exists
     public sendMessageToWebview(debugCommand: string, state: Object) {
         if (this.currentWebviewTarget) {
-            console.log(`Sending message ${debugCommand}`);
             this.currentWebviewTarget.postMessage({ command: debugCommand });
-        } else {
-            console.log(`The webview is not initialized`);
         }
+    }
+    public sendStartMessage() {
+        this.currentWebviewTarget.postMessage({
+            command: VSCODE_MESSAGES_TO_WEBVIEW.RUN_DEVICE,
+        });
+    }
+    public sendStopMessage() {
+        this.currentWebviewTarget.postMessage({
+            command: VSCODE_MESSAGES_TO_WEBVIEW.PAUSE_DEVICE,
+        });
     }
 }
