+++ src/view/App.tsx
@@ -10,20 +10,17 @@ import {
     VIEW_STATE,
 } from "./constants";
 import { Device } from "./container/device/Device";
-import {ViewStateContext} from './context'
+import { ViewStateContext } from "./context";
 
 interface IState {
     currentDevice: string;
-    viewState:VIEW_STATE;
+    viewState: VIEW_STATE;
 }
 
 const defaultState = {
     currentDevice: DEVICE_LIST_KEY.CPX,
-    viewState:VIEW_STATE.RUNNING
+    viewState: VIEW_STATE.RUNNING,
 };
-
-
-
 
 class App extends React.Component<{}, IState> {
     constructor() {
@@ -50,28 +47,30 @@ class App extends React.Component<{}, IState> {
         return (
             <div className="App">
                 <main className="App-main">
-                    <ViewStateContext.Provider value ={this.state.viewState}>
-                    <Device currentSelectedDevice={this.state.currentDevice} />
-                    </ViewStateContext.Provider></main>
+                    <ViewStateContext.Provider value={this.state.viewState}>
+                        <Device
+                            currentSelectedDevice={this.state.currentDevice}
+                        />
+                    </ViewStateContext.Provider>
+                </main>
             </div>
         );
     }
 
     handleMessage = (event: any): void => {
         const message = event.data;
-        console.log(
-            "blabla"+JSON.stringify(message));
+
         switch (message.command) {
             case VSCODE_MESSAGES_TO_WEBVIEW.SET_DEVICE:
                 if (message.active_device !== this.state.currentDevice) {
                     this.setState({ currentDevice: message.active_device });
                 }
                 break;
-            case DEBUG_COMMANDS.CONTINUE:
-                this.setState({viewState:VIEW_STATE.RUNNING})
+            case VSCODE_MESSAGES_TO_WEBVIEW.RUN_DEVICE:
+                this.setState({ viewState: VIEW_STATE.RUNNING });
                 break;
-            case DEBUG_COMMANDS.STACK_TRACE:
-                this.setState({viewState:VIEW_STATE.PAUSE})
+            case VSCODE_MESSAGES_TO_WEBVIEW.PAUSE_DEVICE:
+                this.setState({ viewState: VIEW_STATE.PAUSE });
 
                 break;
         }
