+++ src/view/constants.ts
@@ -56,9 +56,9 @@ export enum DEVICE_LIST_KEY {
 }
 
 // Pauses on Debug mode alter the state of the view
-export enum VIEW_STATE{
-    PAUSE="debug-pause",
-    RUNNING="running"
+export enum VIEW_STATE {
+    PAUSE = "debug-pause",
+    RUNNING = "running",
 }
 
 //
@@ -72,6 +72,8 @@ export enum WEBVIEW_MESSAGES {
 }
 export enum VSCODE_MESSAGES_TO_WEBVIEW {
     SET_DEVICE = "set-device",
+    PAUSE_DEVICE = "pause-device",
+    RUN_DEVICE = "run-device",
 }
 export enum DEBUG_COMMANDS {
     STACK_TRACE = "stackTrace",
